pyproject.toml
[project]
name = "fortify-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.31.0",
    "urllib3>=2.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = []
package = true

[project.scripts]
fortify = "fortify_tool.main:main"
